Move `install_conda_dependencies` into base package
